revert: restore AWS Lambda routing for /api/demo-auth
Reverts the demo-auth routing changes from PR #523, which:
- Reintroduced a Netlify function that the April 5 migration had already archived
- Violated the cost / compliance policy in .github/copilot-instructions.md
  (HIPAA BAA, FedRAMP, and ~$250/month cost delta per 1M requests)
- Never actually fixed the validation failure because JWT_SECRET was
  not configured on Netlify — every POST returns HTTP 500

Changes:
- netlify.toml: /api/demo-auth -> https://api.securebase.tximhotep.com/demo-auth
- phase3a-portal/netlify.toml: same
- phase3a-portal/netlify/functions/demo-auth.js moved to archived/netlify-functions/
- archived/netlify-functions/MIGRATION_NOTES.md: reversion context appended

The daily validation will remain red until the AWS Lambda HTTP 403 is
fixed — see companion diagnostic issue.
