+++ archived/netlify-functions/MIGRATION_NOTES.md
@@ -62,3 +62,39 @@
 - CloudTrail audit integration
 - Single-cloud narrative (AWS-only stack)
 - Consistent credential management via AWS Secrets Manager
+
+---
+
+## demo-auth.js — Re-archival after PR #523 reversion (April 2026)
+
+### Context
+PR #523 (merged 2026-04-15 while still marked `[WIP]`) reintroduced a
+Netlify function at `phase3a-portal/netlify/functions/demo-auth.js` and
+rerouted `/api/demo-auth` away from the AWS Lambda and back to Netlify.
+This reversed the April 5 migration above and violated the cost /
+compliance policy in `.github/copilot-instructions.md` (Cost Optimization:
+Avoid Netlify Functions).
+
+Daily validation (issue #524) continued to fail because the new Netlify
+function required `JWT_SECRET` or `RSA_PRIVATE_KEY` which was never set
+on the Netlify site, so every login attempt returned HTTP 500.
+
+### Reversion
+- Both `netlify.toml` files restored: `/api/demo-auth` → `https://api.securebase.tximhotep.com/demo-auth`
+- `phase3a-portal/netlify/functions/demo-auth.js` re-archived to this directory
+- Overwrites the April 5 archival with the more complete 5-account implementation from PR #523 for reference only — it MUST NOT be re-deployed as a Netlify function.
+
+### Remaining work
+The AWS Lambda itself was returning HTTP 403 at the time PR #523 was opened
+(see PR #523 description). That root cause was never diagnosed — PR #523
+sidestepped it by introducing a second, unauthorised implementation.
+Tracked in companion diagnostic issue filed alongside the reversion PR.
+
+### Policy reminder
+Per `.github/copilot-instructions.md`:
+> Do NOT introduce new Netlify Functions. Existing functions are
+> grandfathered but should be migrated opportunistically.
+
+`demo-auth` was already migrated. Any future change to this endpoint MUST
+be made on the AWS Lambda (`landing-zone/modules/demo-backend/lambda/auth.py`)
+and exposed via the existing netlify.toml redirect.
